requirements.txt
google
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ipywidgets
jupyter
jupyterlab
jupyterlab-vim
numpy
pandas
pandera
psycopg2
pyspark
pytest
pyyaml
seaborn
streamlit
sqlalchemy
tqdm
# pyreadr
# rpy2
